actualizando borrado automatico

+++ app/src/main/java/com/example/comprasmu/ui/mantenimiento/BorrarDatosViewModel.java
@@ -1,63 +1,64 @@
 package com.example.comprasmu.ui.mantenimiento;
 
-import android.app.Activity;
+
 import android.app.Application;
-import android.app.ListActivity;
 import android.content.Context;
 import android.util.Log;
 
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModel;
-
 import com.example.comprasmu.data.ComprasDataBase;
 import com.example.comprasmu.data.dao.ListaCompraDao;
+import com.example.comprasmu.data.modelos.DetalleCaja;
 import com.example.comprasmu.data.modelos.ImagenDetalle;
 import com.example.comprasmu.data.modelos.InformeCompra;
 import com.example.comprasmu.data.modelos.InformeCompraDetalle;
-import com.example.comprasmu.data.modelos.InformeWithDetalle;
+import com.example.comprasmu.data.modelos.InformeEtapa;
 import com.example.comprasmu.data.modelos.ListaCompra;
 import com.example.comprasmu.data.modelos.ListaCompraDetalle;
-import com.example.comprasmu.data.modelos.ProductoExhibido;
 import com.example.comprasmu.data.modelos.Visita;
 import com.example.comprasmu.data.modelos.VisitaWithInformes;
 import com.example.comprasmu.data.repositories.ImagenDetRepositoryImpl;
+import com.example.comprasmu.data.repositories.InfEtapaDetRepoImpl;
+import com.example.comprasmu.data.repositories.InfEtapaRepositoryImpl;
 import com.example.comprasmu.data.repositories.InformeComDetRepositoryImpl;
 import com.example.comprasmu.data.repositories.InformeCompraRepositoryImpl;
 import com.example.comprasmu.data.repositories.ListaCompraDetRepositoryImpl;
 import com.example.comprasmu.data.repositories.ListaCompraRepositoryImpl;
 import com.example.comprasmu.data.repositories.ProductoExhibidoRepositoryImpl;
 import com.example.comprasmu.data.repositories.VisitaRepositoryImpl;
-import com.example.comprasmu.utils.EliminadorIndice;
+import com.example.comprasmu.utils.ComprasLog;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Observable;
+
 
 public class BorrarDatosViewModel extends AndroidViewModel {
 
-
-
-
- /*   InformeComDetRepositoryImpl icdrepo;
+    InformeComDetRepositoryImpl icdrepo;
     InformeCompraRepositoryImpl icrepo;
-    ImagenDetRepositoryImpl imrepo;*/
+    InfEtapaRepositoryImpl ieRepo;
+    InfEtapaDetRepoImpl iedRepo;
+    DetalleCaja detCaja;
+    ImagenDetRepositoryImpl imrepo;
     ListaCompraRepositoryImpl lcrepo;
     ListaCompraDetRepositoryImpl lcdrepo;
-  /*  ProductoExhibidoRepositoryImpl prorepo;
-    VisitaRepositoryImpl visitarepo;*/
+    ProductoExhibidoRepositoryImpl prorepo;
+    VisitaRepositoryImpl visitarepo;
 
     Context context;
+    ComprasLog complog;
     public BorrarDatosViewModel(Application application) {
         super(application);
         this.context = application;
+        complog=ComprasLog.getSingleton();
 
         lcdrepo = new ListaCompraDetRepositoryImpl(application);
     }
 
-   /* File carpeta;
+    File carpeta;
 
     public void setCarpeta(File carpeta) {
         this.carpeta = carpeta;
@@ -89,7 +90,7 @@ public class BorrarDatosViewModel extends AndroidViewModel {
                 //borro los detalle
                     icdrepo.deleteByInforme(informe.getId());
                 //busco imagenes del informe
-                borrarImagenes(visita.visita,informe);
+                      borrarImagenes(visita.visita,informe);
                     icrepo.deleteInformeCompra(informe.getId());
                     } //borre el informe
                 }
@@ -112,35 +113,34 @@ public class BorrarDatosViewModel extends AndroidViewModel {
 
     }
 
+    public void borrarInformesetapa(String indice){
+        ieRepo=new InfEtapaRepositoryImpl(context);
+        iedRepo=new InfEtapaDetRepoImpl(context);
+        complog.grabarError("borrando informes etapas");
+        ieRepo.deleteByIndice(indice);
+        iedRepo.deleteAll();
+
+    }
+
 
 
     private void puedoBorrarInforme(InformeCompra informe){
 
-        icdrepo.getAll(informe.getId()).observeForever(new Observer<List<InformeCompraDetalle>>() {
-               @Override
-               public void onChanged(List<InformeCompraDetalle> informeCompraDetalles) {
-                   if(informeCompraDetalles==null||informeCompraDetalles.size()==0){
-                    //TODO hay que cambiar el proceso
-                       //reviso la imagenes
-                     /*  imrepo.getAllByInforme(informe.getId()).observeForever(new Observer<List<ImagenDetalle>>() {
-                                  @Override
-                                  public void onChanged(List<ImagenDetalle> imagenDetalles) {
-                                      if(informeCompraDetalles==null||informeCompraDetalles.size()==0){
-                                            //puedo borrar
-                                         icrepo.deleteInformeCompra(informe.getId());
+        List<InformeCompraDetalle> informeCompraDetalles=icdrepo.getAllSencillo(informe.getId());
+        if(informeCompraDetalles==null||informeCompraDetalles.size()==0) {
 
-                                      }
+            //reviso la imagenes
+            List<ImagenDetalle> imagenDetalles = imrepo.getFotosInf(informe);
 
-                                  }
-                              }
-                       );*/
-             /*      }
+            if (imagenDetalles == null || imagenDetalles.size() == 0) {
+                //puedo borrar si ya no hay detalles ni imagenes
+                icrepo.deleteInformeCompra(informe.getId());
 
-               }
-           }
-        );
+            }
+        }
 
-    }*/
+
+    }
 
     private void puedoBorrarLista(ListaCompra compra){
 
@@ -179,7 +179,7 @@ public class BorrarDatosViewModel extends AndroidViewModel {
                 }
 
             }
-  /*  public void borrarImagenes(Visita visita, InformeCompra informe){
+    public void borrarImagenes(Visita visita, InformeCompra informe){
         //todas las fotos aqui
         List<Integer> fotosinfo=new ArrayList<>();
         //la visita
@@ -200,6 +200,6 @@ public class BorrarDatosViewModel extends AndroidViewModel {
             if (fdelete.delete()) {
                 System.out.println("file Deleted :" + path); }
             else { Log.e("BorrarDatosFregment","No se pudo borrar el archivo "+path); } }
-    }*/
+    }
 
 }
